\name{bf.test}
\alias{bf.test}
\title{BayesFactor test for BayesTraits models.}
\usage{
bf.test(model1, model2)
}
\description{
	Perform Bayes factor test for output from a pair of BayesTraits Bayesian models. 
}
\arguments{
  \item{model1}{output from a Bayesian BayesTraits analysis.}
  \item{model2}{output from a Bayesian BayesTraits analysis.}
}
\details{
}
\value{
	Results from Bayes factor test in a dataframe.
}
\author{
	Randi H. Griffin 
}
\references{
	Bernardo, J., Smith, A.F.M. (1994). "Bayesian Theory". John Wiley.
}


