% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw-package.R
\docType{package}
\name{btw-package}
\alias{btw-package}
\title{btw: Describe R Stuff to Large Language Models}
\description{
Provides a number of utilities for describing R objects and package documentation in plain text. For interactive use, this is especially powerful for describing relevant pieces of context to large language models. When used programmatically, these utilities can be registered with 'ellmer' chats as tool calls, enabling language models to peruse package documentation and explore your computational environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/posit-dev/btw}
  \item \url{https://posit-dev.github.io/btw/}
  \item Report bugs at \url{https://github.com/posit-dev/btw/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
