% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-agent-researcher-r-task.R
\name{btw_agent_researcher_r_task}
\alias{btw_agent_researcher_r_task}
\title{Agent: A Tool for Researching R Tasks}
\usage{
btw_agent_researcher_r_task(
  client = btw_client(tools = FALSE),
  tools = c("docs", "search", "session")
)
}
\arguments{
\item{client}{An \link[ellmer:Chat]{ellmer::Chat} client, defaults to \code{\link[=btw_client]{btw_client()}}.}

\item{tools}{A list of tools to use with the agent. See \code{\link[=btw_client]{btw_client()}} for
details on how to specify tools or \code{\link[=btw_tools]{btw_tools()}} for the list of built-in
tools. If provided, \code{tools} replaces the set of tools in \code{client}; if
\code{NULL} the \code{client} retains its existing registered tools.}
}
\value{
Returns an aync \code{\link[ellmer:tool]{ellmer::tool()}} that can be used to invoke the
R task research agent.
}
\description{
An agent that researches R packages and provides example code for specific R
tasks. This agent is designed to help users find the best R packages for
specific tasks, list key functions and arguments, and return runnable
example code.
}
\seealso{
Other agents: 
\code{\link{btw_tool_agent}()}
}
\concept{agents}
