% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_btw_init.R
\name{btw_task_btw_init}
\alias{btw_task_btw_init}
\title{Task: Initialize Project Context File}
\usage{
btw_task_btw_init(
  ...,
  path = "btw.md",
  client = NULL,
  mode = c("app", "console", "client", "tool")
)
}
\arguments{
\item{...}{Additional context to provide to the AI. This can be any text or
R objects that can be converted to text using \code{\link[=btw]{btw()}}.}

\item{path}{The path to the context file to create. Defaults to \code{btw.md}.}

\item{client}{An \link[ellmer:Chat]{ellmer::Chat} client or a \code{provider/model} string to be
passed to \code{\link[ellmer:chat-any]{ellmer::chat()}} to create a chat client. Defaults to
\code{\link[ellmer:chat_anthropic]{ellmer::chat_anthropic()}}. You can use the \code{btw.client} option to set a
default client for new \code{btw_client()} calls, or use a \code{btw.md} project file
for default chat client settings, like provider and model. We check the
\code{client} argument, then the \code{btw.client} R option, and finally the \code{btw.md}
project file, using only the client definition from the first of these that
is available.}

\item{mode}{The mode to run the task in, which affects what is returned from
this function. \code{"app"} and \code{"console"} modes launch interactive sessions,
while \code{"client"} and \code{"tool"} modes return objects for programmatic use.}
}
\value{
When \code{mode} is \code{"app"} or \code{"console"}, this function launches an
interactive session in the browser or the R console, respectively. The
ellmer chat object with the conversation history is returned invisibly
when the session ends.

When \code{mode} is \code{"client"}, this function returns the configured ellmer
chat client object. When \code{mode} is \code{"tool"}, this function returns an
ellmer tool object that can be used in other chat instances.
}
\description{
Create a comprehensive context \code{btw.md} or \code{AGENTS.md} file for your project.
If launched in app or console mode, this task will start an interactive chat
session to guide you through the process of creating a context file.

This task focuses on documenting project context for developers and agents.
See \code{\link[=btw_client]{btw_client()}} for additional details about the format and usage of the
\code{btw.md} context file, including choosing the default LLM provider and model
or the default set of tools to use with \code{\link[=btw_client]{btw_client()}}.
}
\examples{
btw_task_btw_init(mode = "tool")

}
\concept{task and agent functions}
