% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{btw_this.data.frame}
\alias{btw_this.data.frame}
\alias{btw_this.tbl}
\title{Describe a data frame in plain text}
\usage{
\method{btw_this}{data.frame}(
  x,
  ...,
  format = c("skim", "glimpse", "print", "json"),
  dims = c(5, 100)
)

\method{btw_this}{tbl}(
  x,
  ...,
  format = c("skim", "glimpse", "print", "json"),
  dims = c(5, 100)
)
}
\arguments{
\item{format}{One of \code{"skim"}, \code{"glimpse"}, \code{"print"}, or \code{"json"}.
\itemize{
\item \code{"skim"} is the most information-dense format for describing the data. It
uses and returns the same information as \code{\link[skimr:skim]{skimr::skim()}} but formatting as
a JSON object that describes the dataset.
\item To glimpse the data column-by-column, use \code{"glimpse"}. This is
particularly helpful for getting a sense of data frame column names,
types, and distributions, when pairings of entries in individual rows
aren't particularly important.
\item To just print out the data frame, use \code{print()}.
\item To get a json representation of the data, use \code{"json"}. This is
particularly helpful when the pairings among entries in specific rows
are important to demonstrate.
}}

\item{dims}{The number of rows and columns to show, as a numeric vector of
length two. For example, the default \code{dims = c(5, 100)} shows the first 5
rows and 100 columns, whereas \code{dims = c(Inf, Inf)} would show all of the data.}

\item{data_frame}{A single string naming a data frame or, alternatively,
the data frame itself.}
}
\value{
A character vector containing a representation of the data frame.
Will error if the named data frame is not found in the environment.
}
\description{
Describe a data frame in plain text
}
\section{Functions}{
\itemize{
\item \code{btw_this(data.frame)}: Summarize a data frame.

\item \code{btw_this(tbl)}: Summarize a \code{tbl}.

}}
\section{See Also}{

\itemize{
\item \code{\link[=register_btw_tools]{register_btw_tools()}} registers this function as a tool for LLMs to call.
}
}

\examples{
btw_this(mtcars)

btw_this(mtcars, format = "print")

btw_this(mtcars, format = "json")

}
