% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-env.R
\name{btw_this.env}
\alias{btw_this.env}
\title{Describe the contents of an environment}
\usage{
\method{btw_this}{env}(x, ..., items = NULL)
}
\arguments{
\item{items}{Optional. A character vector of objects in the environment to
describe.}

\item{environment}{An environment. Optional, defaults to global environment.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Describe the contents of an environment
}
\section{See Also}{

\itemize{
\item \code{\link[=register_btw_tools]{register_btw_tools()}} registers this function as a tool for LLMs to call.
\item Other \verb{get_*()} functions: \code{\link[=get_data_frame]{get_data_frame()}}, \code{\link[=btw_describe_environment]{btw_describe_environment()}}, \code{\link[=get_installed_packages]{get_installed_packages()}}, \code{\link[=get_package_help]{get_package_help()}}, \code{\link[=get_package_vignette]{get_package_vignette()}}, \code{\link[=get_package_vignettes]{get_package_vignettes()}}, \code{\link[=get_help_page]{get_help_page()}}
}
}

