% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-env.R
\name{btw_this.environment}
\alias{btw_this.environment}
\title{Describe the contents of an environment}
\usage{
\method{btw_this}{environment}(x, ..., items = NULL)
}
\arguments{
\item{items}{Optional. A character vector of objects in the environment to
describe.}

\item{environment}{An environment. Optional, defaults to global environment.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Describe the contents of an environment
}
\section{See Also}{

\itemize{
\item \code{\link[=register_btw_tools]{register_btw_tools()}} registers this function as a tool for LLMs to call.
}
}

