% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw_this.R
\name{btw_this}
\alias{btw_this}
\title{Describe something for use by an LLM}
\usage{
btw_this(x, ...)
}
\arguments{
\item{x}{The thing to describe.}

\item{...}{Additional arguments passed down to underlying methods. Unused
arguments are silently ignored.}
}
\value{
A character vector of lines describing the object.
}
\description{
A generic function used to describe an object for use by LLM.
}
\examples{
btw_this(mtcars) # describe the mtcars dataset
btw_this(dplyr::mutate) # include function source
\dontrun{
btw_this("{dplyr}") # include dplyr's intro vignette
btw_this("./") # list files in the current working directory
}

}
\seealso{
Other btw formatting methods: 
\code{\link{btw_this.character}()},
\code{\link{btw_this.data.frame}()},
\code{\link{btw_this.environment}()}
}
\concept{btw formatting methods}
