% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-agent.R
\name{btw_tool_agent}
\alias{btw_tool_agent}
\title{Tool: Create An Agent}
\usage{
btw_tool_agent(
  ...,
  client = NULL,
  system_prompt = NULL,
  turns = NULL,
  mirai_profile = "sidechat",
  setup_code = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[mirai:mirai]{mirai::mirai()}}. Can be used to
adjust \code{\link[mirai:mirai]{mirai::mirai()}} options or to pass values from the parent session
when they are used in \code{setup_code}.}

\item{client}{An \link[ellmer:Chat]{ellmer::Chat} client. If not provided, uses \code{btw_client()}
by default.}

\item{system_prompt}{Additional instructions added to the system prompt of
\code{client} that describe the agent's task and how it should operate.}

\item{turns}{A list of \link[ellmer:Turn]{ellmer::Turn}s or an \link[ellmer:Chat]{ellmer::Chat} object that
provides the initial turns used by the \code{client}. For example, you can use
this argument to pass a chat client with a long chat history to ask the
agent to summarize the chat history.}

\item{mirai_profile}{The name of the \pkg{mirai} \code{.compute} profile to use
for this agent, see \code{\link[mirai:mirai]{mirai::mirai()}} for details.}

\item{setup_code}{Additional code to run first before invoking the \verb{$chat()}
method of \code{client}. Use this argument to load required packages, for
example if \code{client} uses tools from custom packages.}
}
\value{
Returns an \code{\link[ellmer:tool]{ellmer::tool()}} that creates an agent
}
\description{
A btw agent is simply an \link[ellmer:Chat]{ellmer::Chat} client wrapped into a tool call.
}
