% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-cran.R
\name{btw_tool_cran_package}
\alias{btw_tool_cran_package}
\title{Tool: Describe a CRAN package}
\usage{
btw_tool_cran_package(package_name, `_intent` = "")
}
\arguments{
\item{package_name}{The name of a package on CRAN.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}

\item{after}{Only return releases on or after this ISO date (\code{YYYY-MM-DD}).}

\item{before}{Only return releases on or before this ISO date
(\code{YYYY-MM-DD}).}
}
\value{
An info sheet about the package.
}
\description{
Describes a CRAN package using \code{\link[pkgsearch:cran_package]{pkgsearch::cran_package()}}.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) withAutoprint(\{ # examplesIf}
cli::cat_line(
  btw_this(pkgsearch::cran_package("anyflights"))
)

\dontshow{\}) # examplesIf}
}
\seealso{
Other cran tools:
\code{\link[=btw_tool_cran_search]{btw_tool_cran_search()}},
\code{\link[=btw_tool_cran_versions]{btw_tool_cran_versions()}}
}
\concept{cran tools}
