% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-cran.R
\name{btw_tool_cran_search}
\alias{btw_tool_cran_search}
\title{Tool: Search for an R package on CRAN}
\usage{
btw_tool_cran_search(
  query,
  format = c("short", "long"),
  n_results = NULL,
  `_intent` = ""
)
}
\arguments{
\item{query}{Search query string. If this argument is missing or
\code{NULL}, then the results of the last query are printed, in
\emph{short} and \emph{long} formats, in turns for successive
\code{pkg_search()} calls. If this argument is missing, then all
other arguments are ignored.}

\item{format}{Default formatting of the results. \emph{short} only
outputs the name and title of the packages, \emph{long} also
prints the author, last version, full description and URLs.
Note that this only affects the default printing, and you can
still inspect the full results, even if you specify \emph{short}
here.}

\item{n_results}{Number of search results to include. Defaults to 10 for
'short' format and 5 for 'long' format.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
A listing of packages matching the search term.
}
\description{
Uses \code{\link[pkgsearch:pkg_search]{pkgsearch::pkg_search()}} to search for R packages on CRAN.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) withAutoprint(\{ # examplesIf}
# Copy pkgsearch results to the clipboard for use in any LLM app
btw(
  pkgsearch::pkg_search("network visualization", size = 1),
  clipboard = FALSE
)
btw(
  pkgsearch::pkg_search("network visualization", format = "long", size = 1),
  clipboard = FALSE
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other cran tools:
\code{\link[=btw_tool_cran_package]{btw_tool_cran_package()}},
\code{\link[=btw_tool_cran_versions]{btw_tool_cran_versions()}}
}
\concept{cran tools}
