% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-sessioninfo.R
\name{btw_tool_describe_platform}
\alias{btw_tool_describe_platform}
\title{Tool: Describe user's platform}
\usage{
btw_tool_describe_platform()
}
\value{
Returns a single string describing the user's platform.
}
\description{
Describes the R version, operating system, and language and locale settings
for the user's system. When using \code{\link[=btw_client]{btw_client()}} or \code{\link[=btw_app]{btw_app()}}, this
information is automatically included in the system prompt.
}
\examples{
cat(btw_tool_describe_platform())

}
\seealso{
\code{\link[=btw_register_tools]{btw_register_tools()}}

Other Tools: 
\code{\link{btw_register_tools}()},
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_describe_environment}()},
\code{\link{btw_tool_get_installed_packages}()},
\code{\link{btw_tool_list_files}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_read_current_editor}()},
\code{\link{btw_tool_read_text_file}()}
}
\concept{Tools}
