% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-query.R
\name{btw_tool_env_query_data_frame}
\alias{btw_tool_env_query_data_frame}
\title{Perform a SQL query on the data, and return the results as JSON.}
\usage{
btw_tool_env_query_data_frame(query, data_frame)
}
\arguments{
\item{query}{A DuckDB SQL query; must be a SELECT statement.}

\item{data_frame}{The name of the data frame.}
}
\value{
The results of the query as a JSON string.
}
\description{
Perform a SQL query on the data, and return the results as JSON.
}
