% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-files-code-search.R
\name{btw_tool_files_code_search}
\alias{btw_tool_files_code_search}
\title{Tool: Code Search in Project}
\usage{
btw_tool_files_code_search(
  term,
  limit = 100,
  case_sensitive = TRUE,
  use_regex = FALSE,
  .intent = ""
)
}
\arguments{
\item{term}{The term to search for in the code files.}

\item{limit}{Maximum number of matching lines to return (between 1 and 1000,
default 100).}

\item{case_sensitive}{Whether the search should be case-sensitive (default is
\code{FALSE}).}

\item{use_regex}{Whether to interpret the search term as a regular expression
(default is \code{FALSE}).}
}
\value{
Returns a tool result with a data frame of search results, with
columns for \code{filename}, \code{size}, \code{last_modified}, \code{content} and \code{line}.
}
\description{
Search through code files in the project directory for specific terms.
}
\seealso{
Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_files_write_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tool_web_read_url}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
