% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-search-web.R
\name{btw_tool_files_search_ddg}
\alias{btw_tool_files_search_ddg}
\title{Tool: Search DuckDuckGo}
\usage{
btw_tool_files_search_ddg(keywords, max_results = 3)
}
\arguments{
\item{keywords}{Search keywords}

\item{max_results}{Maximum number of results to return, between 1 and 10.}
}
\value{
Returns a markdown table of search results.
}
\description{
Tool: Search DuckDuckGo
}
\examples{
btw_tool_files_search_ddg("R programming language", max_results = 3)

}
\seealso{
Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_files_write_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
