% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tool-files.R
\name{btw_tool_files_write_text_file}
\alias{btw_tool_files_write_text_file}
\alias{btw_tool_files_write}
\title{Tool: Write a text file}
\usage{
btw_tool_files_write_text_file(path, content, `_intent` = NULL)

btw_tool_files_write(path, content, `_intent` = "")
}
\arguments{
\item{path}{Path to the file to write. The \code{path} must be in the current
working directory.}

\item{content}{The text content to write to the file. This should be the
complete content as the file will be overwritten.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns a message confirming the file was written.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{btw_tool_files_write_text_file()} was renamed to
\code{\link[=btw_tool_files_write]{btw_tool_files_write()}} in btw 1.2.0.
}
\examples{
withr::with_tempdir({
  btw_tool_files_write("example.txt", "Hello\nWorld!")
  readLines("example.txt")
})

}
\seealso{
Other files tools: 
\code{\link{btw_tool_files_code_search}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()}
}
\concept{files tools}
\keyword{internal}
