% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-docs.R
\name{btw_tool_get_installed_packages}
\alias{btw_tool_get_installed_packages}
\title{Tool: Describe installed packages}
\usage{
btw_tool_get_installed_packages()
}
\value{
Returns a single string describing installed packages as JSON.
}
\description{
Displays the name and title of all installed R packages in JSON format.
}
\examples{
cat(btw_tool_get_installed_packages())

}
\seealso{
\code{\link[=btw_register_tools]{btw_register_tools()}}

Other Tools: 
\code{\link{btw_register_tools}()},
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_describe_environment}()},
\code{\link{btw_tool_describe_platform}()},
\code{\link{btw_tool_list_files}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_read_current_editor}()},
\code{\link{btw_tool_read_text_file}()},
\code{\link{btw_tool_session_package_info}()}
}
\concept{Tools}
