% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-git.R
\name{btw_tool_git_status}
\alias{btw_tool_git_status}
\title{Tool: Git Status}
\usage{
btw_tool_git_status(staged = NULL, pathspec = NULL, `_intent` = "")
}
\arguments{
\item{staged}{Optional. Return only staged (\code{TRUE}) or unstaged files
(\code{FALSE}). Use \code{NULL} to show both (default).}

\item{pathspec}{Optional character vector with paths to match.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns a character table of file statuses.
}
\description{
Tool: Git Status
}
\examples{
\dontrun{
btw_tool_git_status()
}

}
\seealso{
Other git tools: 
\code{\link{btw_tool_git_branch_checkout}()},
\code{\link{btw_tool_git_branch_create}()},
\code{\link{btw_tool_git_branch_list}()},
\code{\link{btw_tool_git_commit}()},
\code{\link{btw_tool_git_diff}()},
\code{\link{btw_tool_git_log}()}
}
\concept{git tools}
