% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-github.R
\name{btw_tool_github_issue_thread}
\alias{btw_tool_github_issue_thread}
\title{Tool: GitHub Issue Thread}
\usage{
btw_tool_github_issue_thread(
  issue_number,
  owner = NULL,
  repo = NULL,
  `_intent` = ""
)
}
\arguments{
\item{issue_number}{The issue number to retrieve.}

\item{owner}{Optional GitHub repository owner.}

\item{repo}{Optional GitHub repository name.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns the issue with all comments.
}
\description{
Tool: GitHub Issue Thread
}
\examples{
\dontrun{
btw_tool_github_issue_thread(issue_number = 123)
}

}
\seealso{
Other github tools: 
\code{\link{btw_tool_github_issue_create}()},
\code{\link{btw_tool_github_issue_get}()},
\code{\link{btw_tool_github_issues_list}()},
\code{\link{btw_tool_github_pull_request_create}()},
\code{\link{btw_tool_github_pull_request_diff}()},
\code{\link{btw_tool_github_pull_request_get}()},
\code{\link{btw_tool_github_pull_requests_list}()}
}
\concept{github tools}
