% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-github.R
\name{btw_tool_github_pull_request_create}
\alias{btw_tool_github_pull_request_create}
\title{Tool: GitHub Pull Request Create}
\usage{
btw_tool_github_pull_request_create(
  title,
  body,
  head,
  base,
  owner = NULL,
  repo = NULL,
  `_intent` = ""
)
}
\arguments{
\item{title}{The pull request title.}

\item{body}{The pull request body/description.}

\item{head}{The name of the branch containing changes.}

\item{base}{The name of the branch to merge into.}

\item{owner}{Optional GitHub repository owner.}

\item{repo}{Optional GitHub repository name.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns the created pull request details.
}
\description{
Tool: GitHub Pull Request Create
}
\examples{
\dontrun{
btw_tool_github_pull_request_create(
  title = "Add new feature",
  body = "This PR adds...",
  head = "feature-branch",
  base = "main"
)
}

}
\seealso{
Other github tools: 
\code{\link{btw_tool_github_issue_create}()},
\code{\link{btw_tool_github_issue_get}()},
\code{\link{btw_tool_github_issue_thread}()},
\code{\link{btw_tool_github_issues_list}()},
\code{\link{btw_tool_github_pull_request_diff}()},
\code{\link{btw_tool_github_pull_request_get}()},
\code{\link{btw_tool_github_pull_requests_list}()}
}
\concept{github tools}
