% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-pkg-devtools.R
\name{btw_tool_pkg_document}
\alias{btw_tool_pkg_document}
\title{Tool: Generate package documentation}
\usage{
btw_tool_pkg_document(pkg = ".", `_intent` = "")
}
\arguments{
\item{pkg}{Path to package directory. Defaults to \code{"."}. Must be within
current working directory.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
The output from \code{\link[devtools:document]{devtools::document()}}.
}
\description{
Generate package documentation using \code{\link[devtools:document]{devtools::document()}}. This runs
\pkg{roxygen2} on the package to create/update man pages and \code{NAMESPACE}.
}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other pkg tools: 
\code{\link{btw_tool_pkg_check}()},
\code{\link{btw_tool_pkg_coverage}()},
\code{\link{btw_tool_pkg_load_all}()},
\code{\link{btw_tool_pkg_test}()}
}
\concept{pkg tools}
