% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-files.R
\name{btw_tool_read_text_file}
\alias{btw_tool_read_text_file}
\title{Tool: Read a file}
\usage{
btw_tool_read_text_file(path, max_lines = 1000)
}
\arguments{
\item{path}{Path to a file for which to get information. The \code{path} must be
in the current working directory.}

\item{max_lines}{Number of lines to include. Defaults to 1,000 lines.}
}
\value{
Returns a character vector of lines from the file.
}
\description{
Tool: Read a file
}
\seealso{
Other Tools: 
\code{\link{btw_register_tools}()},
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_describe_environment}()},
\code{\link{btw_tool_list_files}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_read_current_editor}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()}
}
\concept{Tools}
