\docType{data}
\name{continuous2}
\alias{continuous2}
\title{Random pair of continuous traits for primates}
\format{Text file}
\source{
	Random data generated in R
}
\description{
	This is just random data to demonstrate how to use
	BayesTraits functions.
}
\author{
      Randi Griffin
}
\usage{continuous2}
\keyword{datasets}