% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{btw_tool_session_platform_info}
\alias{btw_tool_session_package_info}
\alias{btw_tool_session_check_package_installed}
\alias{btw_tool_search_packages}
\alias{btw_tool_search_package_info}
\alias{btw_tool_files_list_files}
\alias{btw_tool_files_read_text_file}
\alias{btw_tool_files_write_text_file}
\alias{btw_tool_files_code_search}
\title{Deprecated functions}
\usage{
btw_tool_session_platform_info(`_intent` = NULL)

btw_tool_session_package_info(
  packages = "attached",
  dependencies = "",
  `_intent` = NULL
)

btw_tool_session_check_package_installed(package_name, `_intent` = NULL)

btw_tool_search_packages(
  query,
  format = c("short", "long"),
  n_results = NULL,
  `_intent` = NULL
)

btw_tool_search_package_info(package_name, `_intent` = NULL)

btw_tool_files_list_files(
  path = NULL,
  type = c("any", "file", "directory"),
  regexp = "",
  `_intent` = NULL
)

btw_tool_files_read_text_file(
  path,
  line_start = 1,
  line_end = 1000,
  `_intent` = NULL
)

btw_tool_files_write_text_file(path, content, `_intent` = NULL)

btw_tool_files_code_search(
  term,
  limit = 100,
  case_sensitive = TRUE,
  use_regex = FALSE,
  show_lines = FALSE,
  `_intent` = ""
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been renamed. The old names continue to work but emit
deprecation warnings. Use the new names for new code.
}
\section{Functions}{
\itemize{
\item \code{btw_tool_session_platform_info()}: \code{btw_tool_session_platform_info()} was renamed to
\code{\link[=btw_tool_sessioninfo_platform]{btw_tool_sessioninfo_platform()}} in btw 1.2.0.

\item \code{btw_tool_session_package_info()}: \code{btw_tool_session_package_info()} was renamed to
\code{\link[=btw_tool_sessioninfo_package]{btw_tool_sessioninfo_package()}} in btw 1.2.0.

\item \code{btw_tool_session_check_package_installed()}: \code{btw_tool_session_check_package_installed()} was
renamed to \code{\link[=btw_tool_sessioninfo_is_package_installed]{btw_tool_sessioninfo_is_package_installed()}} in btw 1.2.0.

\item \code{btw_tool_search_packages()}: \code{btw_tool_search_packages()} was renamed to
\code{\link[=btw_tool_cran_search]{btw_tool_cran_search()}} in btw 1.2.0.

\item \code{btw_tool_search_package_info()}: \code{btw_tool_search_package_info()} was renamed to
\code{\link[=btw_tool_cran_package]{btw_tool_cran_package()}} in btw 1.2.0.

\item \code{btw_tool_files_list_files()}: \code{btw_tool_files_list_files()} was renamed to
\code{\link[=btw_tool_files_list]{btw_tool_files_list()}} in btw 1.2.0.

\item \code{btw_tool_files_read_text_file()}: \code{btw_tool_files_read_text_file()} was renamed to
\code{\link[=btw_tool_files_read]{btw_tool_files_read()}} in btw 1.2.0.

\item \code{btw_tool_files_write_text_file()}: \code{btw_tool_files_write_text_file()} was renamed to
\code{\link[=btw_tool_files_write]{btw_tool_files_write()}} in btw 1.2.0.

\item \code{btw_tool_files_code_search()}: \code{btw_tool_files_code_search()} was renamed to
\code{\link[=btw_tool_files_search]{btw_tool_files_search()}} in btw 1.2.0.

}}
\keyword{internal}
