\docType{data}
\name{discrete1}
\alias{discrete1}
\title{Random discrete trait data for primates}
\format{Text file}
\source{
	Random data generated in R
}
\description{
	This is just random data to demonstrate how to use
	BayesTraits functions.
}
\author{
      Randi Griffin
}
\usage{discrete1}
\keyword{datasets}