\docType{data}
\name{discrete2}
\alias{discrete2}
\title{Random pair of discrete traits for primates}
\format{Text file}
\source{
	Random data generated in R
}
\description{
	This is just random data to demonstrate how to use
	BayesTraits functions.
}
\author{
      Randi Griffin
}
\usage{discrete2}
\keyword{datasets}