% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data_frame}
\alias{get_data_frame}
\title{Describe a data frame in plain text}
\usage{
get_data_frame(
  data_frame,
  format = c("glimpse", "print", "json"),
  dims = c(5, 100)
)
}
\arguments{
\item{data_frame}{A single string naming a data frame or, alternatively,
the data frame itself.}

\item{format}{One of \code{"glimpse"}, \code{"print"}, or \code{"json"}.
\itemize{
\item To glimpse the data column-by-column, use \code{"glimpse"}. This is
particularly helpful for getting a sense of data frame column names,
types, and distributions, when pairings of entries in individual rows
aren't particularly important.
\item To just print out the data frame, use \code{print()}.
\item To get a json representation of the data, use \code{"json"}. This is
particularly helpful when the pairings among entries in specific rows
are important to demonstrate.
}}

\item{dims}{The number of rows and columns to show, as a numeric vector of
length two. For example, the default \code{dims = c(5, 100)} shows the first 5
rows and 100 columns, whereas \code{dims = c(Inf, Inf)} would show all of the data.}
}
\value{
A character vector containing a representation of the data frame.
Will error if the named data frame is not found in the environment.
}
\description{
Describe a data frame in plain text
}
