% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-env.R
\name{get_environment}
\alias{get_environment}
\title{Describe the contents of an environment}
\usage{
get_environment(environment = global_env(), items = NULL)
}
\arguments{
\item{environment}{An environment. Optional, defaults to global environment.}

\item{items}{Optional. A character vector of objects in the environment to
describe.}
}
\value{
A character vector describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Describe the contents of an environment
}
