% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{get_help_page}
\alias{get_help_page}
\title{Get help page from package}
\usage{
get_help_page(package_name, topic)
}
\arguments{
\item{package_name}{The exact name of the package, e.g. "shiny"}

\item{topic}{The topic_id or alias of the help page, e.g. "withProgress" or
"incProgress"}
}
\value{
The help page, in plain text format
}
\description{
Get help page from package
}
