% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{get_installed_packages}
\alias{get_installed_packages}
\title{Get installed R packages}
\usage{
get_installed_packages(pattern = NULL)
}
\arguments{
\item{pattern}{Optional regular expression (case insensitive, PCRE syntax)
to filter by. If not provided, all installed packages will be returned.}
}
\value{
A dataframe of installed packages
}
\description{
Get installed R packages
}
