% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{get_package_help}
\alias{get_package_help}
\title{Get available help topics for an R package}
\usage{
get_package_help(package_name)
}
\arguments{
\item{package_name}{The exact name of the package, e.g. "shiny"}
}
\value{
JSON data of help topics, along with their aliases
}
\description{
Get available help topics for an R package
}
