% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{get_pkg}
\alias{get_pkg}
\alias{get_package_help}
\alias{get_help_page}
\alias{get_package_vignettes}
\alias{get_package_vignette}
\title{Describe R package documentation}
\usage{
get_package_help(package_name)

get_help_page(package_name, topic)

get_package_vignettes(package_name)

get_package_vignette(package_name, vignette = package_name)
}
\arguments{
\item{package_name}{The name of the package as a string, e.g. \code{"shiny"}.}

\item{topic}{The \code{topic_id} or \code{alias} of the help page, e.g. \code{"withProgress"}
or \code{"incProgress"}. Find \code{topic_id}s or \code{alias}es using \code{get_package_help()}.}

\item{vignette}{The name (or index) of the vignette to retrieve. Defaults to
the "intro" vignette to the package (by the same rules as pkgdown.)}
}
\value{
\itemize{
\item \code{get_package_help()} returns the \code{topic_id}, \code{title}, and \code{aliases} fields
for every topic in a package's documentation as a json-formatted string.
\item \code{get_help_page()} return the help-page for a package topic as a string.
}
}
\description{
These functions describe package documentation in plain text:
}
\section{See Also}{

\itemize{
\item \code{\link[=register_btw_tools]{register_btw_tools()}} registers this function as a tool for LLMs to call.
}
}

\examples{
cat(get_package_help("btw"))

cat(get_help_page("btw", "btw"))

# show the TOC of vignettes in the dplyr package
cat(get_package_vignettes("dplyr"))

# returns a whole bunch of output and relies on
# dplyr to have the mentioned vignettes available
\dontrun{
# grab the intro vignette
cat(get_package_vignette("dplyr"))

# grab the programming vignette specifically
cat(get_package_vignette("dplyr", "programming"))
}
}
