% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{install_btw_cli}
\alias{install_btw_cli}
\title{Install the btw CLI}
\usage{
install_btw_cli(destdir = NULL)
}
\arguments{
\item{destdir}{Directory where the CLI launcher will be installed. If \code{NULL},
the default location used by \code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}} is used.}
}
\value{
The result of \code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}}, invisibly.
}
\description{
Installs the \code{btw} CLI launcher using \code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}}.
\href{https://github.com/r-lib/Rapp}{Rapp} is required to build and install the
CLI.
}
