\name{kill.bt}
\alias{kill.bt}
\title{Kill all BayesTraits processes on computer.}
\usage{
kill.bt()
}
\description{
	Kill all BayesTraits processes on computer.
}
\details{
	Escaping from Continuous or Discrete within R does not mean that the BayesTraits analysis has stopped on your computer, and analyses running in the background can really slow things down. If you start a long BayesTraits process and want to quit it, first hit ESC or Crtl-z to quit the function within R, and then run kill.bt() to stop the processes from running in the background on your computer.
}
\author{
	Randi H. Griffin 
}



