% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-run.R
\name{merge_adjacent_content}
\alias{merge_adjacent_content}
\title{Merge adjacent content of the same type}
\usage{
merge_adjacent_content(contents)
}
\arguments{
\item{contents}{List of Content objects}
}
\value{
List of Content objects with adjacent same-type elements merged
}
\description{
Reduces a list of Content objects by concatenating adjacent elements
of the same mergeable type (ContentCode, ContentMessage, ContentWarning,
ContentError) into single elements.
}
\keyword{internal}
