\name{plotdiscrete}
\alias{plotdiscrete}
\title{Plot Discrete}
\usage{
plotdiscrete(model)
}
\description{
	Plot a transition rate matrix for output from a Discrete BayesTraits analysis of dependent/independent evolution of a pair of discrete characters. 
}
\arguments{
  \item{model}{output from a Discrete BayesTraits analysis.}
}
\value{
	A graphical depiction of the model transition rate matrix.
}
\author{
	Randi H. Griffin 
}



