\docType{data}
\name{primate.continuous1}
\alias{primate.continuous1}
\title{Random continuous trait data for primates}
\format{Text file}
\source{
	Random data generated in R
}
\description{
	This is just random data to demonstrate how to use
	BayesTraits functions.
}
\author{
      Randi H. Griffin
}
\usage{primate.continuous1}
\keyword{datasets}
