\docType{data}
\name{primate.continuous3}
\alias{primate.continuous3}
\title{Random set of three continuous traits for primates}
\format{Text file}
\source{
	Random data generated in R
}
\description{
	This is just random data to demonstrate how to use
	BayesTraits functions.
}
\author{
      Randi H. Griffin
}
\usage{primate.continuous3}
\keyword{datasets}