\docType{data}
\name{primate.tree1}
\alias{primate.tree1}
\title{Single primate phylogeny.}
\format{Nexus}
\source{
	10kTrees Version 3.
}
\description{
	Single primate phylogeny.
}
\author{
      Randi H. Griffin
}
\references{
	Arnold C, Matthews LJ, Nunn CL. 2010. The 10kTrees Website: a new online resource for primate phylogeny, v3. Evol Anthropol 19: 114–118.
}
\usage{primate.tree1}
\keyword{datasets}
