\docType{data}
\name{primate.tree100}
\alias{primate.tree100}
\title{Block of 100 primate phylogenies.}
\format{Nexus}
\source{
	10kTrees Version 3.
}
\description{
	Block of 100 primate phylogenies.
}
\author{
      Randi H. Griffin
}
\references{
	Arnold C, Matthews LJ, Nunn CL. 2010. The 10kTrees Website: a new online resource for primate phylogeny, v3. Evol Anthropol 19: 114–118.
}
\usage{primate.tree100}
\keyword{datasets}
