\docType{data}
\name{primatetree100}
\alias{primatetree100}
\title{Block of 100 primate phylogenies.}
\format{Text file}
\source{
	10kTrees Version 3.
}
\description{
	Block of 100 primate phylogenies.
}
\author{
      Randi H. Griffin
}
\references{
	Arnold C, Matthews LJ, Nunn CL. 2010. The 10kTrees Website: a new online resource for primate phylogeny, v3. Evol Anthropol 19: 114–118.
}
\usage{primatetree100}
\keyword{datasets}
