\name{rjmodels}
\alias{rjmodels}
\title{Overview of models sampled from reversible jump analysis.}
\usage{
rjmodels(model)
}
\description{
	View the total number of models and posterior probability of the top 10 models. 
}
\arguments{
  \item{model}{output from a reversible jump BayesTraits analysis.}
}
\value{
	List of length 2. The first element of the list is the total number of models, and the second element is a table displaying the posterior probability of the top 10 models.
}
\author{
	Randi H. Griffin 
}



