% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw-package.R
\docType{package}
\name{btw-package}
\alias{btw-package}
\title{btw: Describe R Stuff in Plain Text}
\description{
Describes R objects in plain text.
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Garrick Aden-Buie \email{garrick@adenbuie.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
