% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw.R
\name{btw}
\alias{btw}
\title{Plain-text descriptions of R objects}
\usage{
btw(..., clipboard = interactive())
}
\arguments{
\item{...}{Character vectors or, more likely, outputs of \verb{this_*()}
functions like \code{\link[=this_env]{this_env()}}, \code{\link[=this_fn]{this_fn()}}, \code{\link[=this_pkg]{this_pkg()}}, or \code{\link[=this_df]{this_df()}}.
If nothing is provided in this argument, \code{btw()} will just describe the
objects in your global R environment.}

\item{clipboard}{Whether to write the results to the clipboard.
A single logical value; will default to \code{TRUE} when run interactively.}
}
\value{
The combined elements as a string, invisibly. If \code{clipboard} is \code{TRUE}, the
result is also written to the system clipboard.
}
\description{
A short description...
}
