% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw.R
\name{btw}
\alias{btw}
\title{Plain-text descriptions of R objects}
\usage{
btw(..., clipboard = TRUE)
}
\arguments{
\item{...}{Objects to describe from your R environment. You can pass objects
themselves, like data frames or functions, or the function also accepts
output from \verb{btw_tool_*()} functions like
\code{\link[=btw_tool_get_package_help_topics]{btw_tool_get_package_help_topics()}}, \code{\link[=btw_tool_get_help_page]{btw_tool_get_help_page()}}, etc. If
omitted, this function will just describe the elements in your global R
environment.}

\item{clipboard}{Whether to write the results to the clipboard.
A single logical value; will default to \code{TRUE} when run interactively.}
}
\value{
The combined elements as a string, invisibly. If \code{clipboard} is \code{TRUE}, the
result is also written to the system clipboard.
}
\description{
This function allows you to quickly describe your computational environment
to a model by concatenating plain-text descriptions of "R stuff", from
data frames to packages to function documentation.
}
\examples{
btw()

btw(mtcars)

btw(btw::btw)

}
