% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw_client.R
\name{btw_client}
\alias{btw_client}
\alias{btw_app}
\title{Create a btw-enhanced ellmer chat client}
\usage{
btw_client(..., client = NULL, include = NULL)

btw_app(..., client = NULL, include = NULL)
}
\arguments{
\item{...}{Objects and documentation to be included as context in the chat,
via \code{\link[=btw]{btw()}}.}

\item{client}{An \link[ellmer:Chat]{ellmer::Chat} client, defaults to \code{\link[ellmer:chat_claude]{ellmer::chat_claude()}}.
You can use the \code{btw.chat_client} option to set a default client for new
\code{btw_client()} calls.}

\item{include}{Names of tools or tool groups to include when registering
tools, e.g. \code{include = "docs"} to include only the documentation related
tools, or \code{include = c("data", "docs", "environment")}, etc.}
}
\value{
Returns an \link[ellmer:Chat]{ellmer::Chat} object with additional tools registered by
\code{\link[=btw_register_tools]{btw_register_tools()}}. \code{btw_app()} returns the chat object invisibly, and
the chat object with the messages added during the chat session.
}
\description{
Creates an \link[ellmer:Chat]{ellmer::Chat} client, enhanced with the tools from
\code{\link[=btw_register_tools]{btw_register_tools()}}. Use \code{btw_chat()} to create the chat client for
general or interactive use at the console, or \code{btw_app()} to create a chat
client and launch a Shiny app for chatting with a btw-enhanced LLM in your
local workspace.
}
\section{Functions}{
\itemize{
\item \code{btw_client()}: Create a btw-enhanced \link[ellmer:Chat]{ellmer::Chat} client

\item \code{btw_app()}: Create a btw-enhanced client and launch a Shiny app to
chat

}}
\examples{
if (interactive()) {
  withr::local_options(list(
    btw.chat_client = ellmer::chat_ollama(model="llama3.1:8b")
  ))

  btw_chat <- btw_client()
  chat$chat("How can I replace `stop()` calls with functions from the cli package?")
}

}
