% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{btw_register_tools}
\alias{btw_register_tools}
\title{Tools: Register tools from btw}
\usage{
btw_register_tools(chat)
}
\arguments{
\item{chat}{An ellmer \code{Chat} object.}
}
\value{
The chat object with tools registered.
}
\description{
The \code{btw_register_tools()} function equips an ellmer chat to interface with
your computational environment. Chats returned by this function have access
to the tools:
\itemize{
\item \strong{btw_tool_describe_data_frame}: Show the data frame or table or get information about the structure of a data frame or table.
\item \strong{btw_tool_describe_environment}: List and describe items in an environment.
\item \strong{btw_tool_get_installed_packages}: Displays the name and title of all installed R packages in json format.
\item \strong{btw_tool_get_package_help_topics}: Get available help topics for an R package.
\item \strong{btw_tool_get_help_page}: Get help page from package.
\item \strong{btw_tool_get_available_vignettes_in_package}: Get a table of available vignettes for an R package.
\item \strong{btw_tool_get_vignette_from_package}: Get a package vignette in plain text..
}
}
\examples{
# requires an ANTHROPIC_API_KEY
\dontrun{
ch <- ellmer::chat_claude()

btw_register_tools(ch)
}

}
\seealso{
Other Tools: 
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_describe_environment}()},
\code{\link{btw_tool_get_installed_packages}()},
\code{\link{btw_tool_package_docs}}
}
\concept{Tools}
