% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw_this.R
\name{btw_this.character}
\alias{btw_this.character}
\title{Describe objects}
\usage{
\method{btw_this}{character}(x, ..., caller_env = parent.frame())
}
\arguments{
\item{x}{A character string}

\item{...}{Ignored.}

\item{caller_env}{The caller environment.}
}
\value{
A character vector of lines describing the object.
}
\description{
Character strings in \code{btw_this()} are used as shortcuts to many underlying
methods. \code{btw_this()} detects specific formats in the input string to
determine which method to call, or by default it will try to evaluate the
character string as R code and return the appropriate object description.

\code{btw_this()} knows about the following special character string formats:
\itemize{
\item \code{"./path"} \cr
Any string starting with \verb{./} is treated as a relative path.
If the path is a file, we call \code{\link[=btw_tool_read_text_file]{btw_tool_read_text_file()}} and if the path
is a directory we call \code{\link[=btw_tool_list_files]{btw_tool_list_files()}} on the path.
\itemize{
\item \code{btw_this("./data")} lists the files in \verb{data/}.
\item \code{btw_this("./R/load_data.R")} reads the source of the \code{R/load_data.R}
file.
}
\item \code{"{pkgName}"} \cr
A package name wrapped in braces. Returns either the
introductory vignette for the package
(\code{\link[=btw_tool_get_vignette_from_package]{btw_tool_get_vignette_from_package()}}) or a list of help topics if no
such vignette exists (\code{\link[=btw_tool_get_package_help_topics]{btw_tool_get_package_help_topics()}}).
\itemize{
\item \code{btw_this("{dplyr}")} includes dplyr's introductory vignette.
\item \code{btw_this("{btw}")} returns the package help index (because \code{btw}
doesn't have an intro vignette, yet).
}
\item \code{"?help_topic"} \cr
When the string starts with \verb{?}, btw searches R's help
topics using \code{\link[=btw_tool_get_help_page]{btw_tool_get_help_page()}}.
\itemize{
\item \code{btw_this("?dplyr::across")} includes the reference page for
\code{dplyr::across}.
}
}
}
\seealso{
Other \code{btw_this()} methods: 
\code{\link{btw_this}()},
\code{\link{btw_this.data.frame}()},
\code{\link{btw_this.environment}()}
}
\concept{\code{btw_this()} methods}
