% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-env.R
\name{btw_this.environment}
\alias{btw_this.environment}
\title{Describe the contents of an environment}
\usage{
\method{btw_this}{environment}(x, ..., items = NULL)
}
\arguments{
\item{x}{An environment.}

\item{...}{Additional arguments are silently ignored.}

\item{items}{Optional. A character vector of objects in the environment to
describe.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Describe the contents of an environment
}
\examples{
btw_this(globalenv())

}
\seealso{
\code{\link[=btw_tool_describe_environment]{btw_tool_describe_environment()}}
}
