% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-environment.R
\name{btw_this.environment}
\alias{btw_this.environment}
\title{Describe the contents of an environment}
\usage{
\method{btw_this}{environment}(x, ..., items = NULL)
}
\arguments{
\item{x}{An environment.}

\item{...}{Additional arguments are silently ignored.}

\item{items}{Optional. A character vector of objects in the environment to
describe.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Describe the contents of an environment
}
\examples{
cyl_6 <- mtcars[mtcars$cyl == 6, ]
gear_5 <- mtcars[mtcars$gear == 5, ]
btw_this(environment())

}
\seealso{
\code{\link[=btw_tool_env_describe_environment]{btw_tool_env_describe_environment()}}

Other \code{btw_this()} methods: 
\code{\link{btw_this}()},
\code{\link{btw_this.character}()},
\code{\link{btw_this.data.frame}()}
}
\concept{\code{btw_this()} methods}
