% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-agent.R
\name{btw_tool_agent}
\alias{btw_tool_agent}
\title{Tool: Create An Agent}
\usage{
btw_tool_agent(
  name,
  ...,
  description = "",
  title = NULL,
  system_prompt = NULL,
  client = NULL,
  turns = NULL,
  tools = NULL,
  mirai_profile = "btw_agent",
  setup_code = NULL
)
}
\arguments{
\item{name}{The name of the agent tool. This is used to identify the tool
in the chat client or in the tool registry.}

\item{...}{Ignored.}

\item{description}{A description of the agent tool, describing how and when
it should be used. This provides context to the chat client that will
invoke the agent.}

\item{title}{The title of the agent tool, used when displaying the tool
in a tool listing or when showing the agent tool's response.}

\item{system_prompt}{Additional instructions added to the system prompt of
\code{client} that describe the agent's task and how it should operate.}

\item{client}{An \link[ellmer:Chat]{ellmer::Chat} client. If not provided, uses \code{btw_client()}
by default.}

\item{turns}{A list of \link[ellmer:Turn]{ellmer::Turn}s or an \link[ellmer:Chat]{ellmer::Chat} object that
provides the initial turns used by the \code{client}. For example, you can use
this argument to pass a chat client with a long chat history to ask the
agent to summarize the chat history.}

\item{tools}{A list of tools to use with the agent. See \code{\link[=btw_client]{btw_client()}} for
details on how to specify tools or \code{\link[=btw_tools]{btw_tools()}} for the list of built-in
tools. If provided, \code{tools} replaces the set of tools in \code{client}; if
\code{NULL} the \code{client} retains its existing registered tools.}

\item{mirai_profile}{The name of the \pkg{mirai} \code{.compute} profile to use
for this agent, see \code{\link[mirai:mirai]{mirai::mirai()}} for details.}

\item{setup_code}{Additional code to run first before invoking the \verb{$chat()}
method of \code{client}. Use this argument to load required packages, for
example if \code{client} uses tools from custom packages.}
}
\value{
Returns an agent as an \code{\link[ellmer:tool]{ellmer::tool()}}.
}
\description{
A btw agent is simply an \link[ellmer:Chat]{ellmer::Chat} client wrapped into a tool call.
}
\seealso{
Other agents: 
\code{\link{btw_agent_researcher_r_task}()}
}
\concept{agents}
