% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tool-cran.R
\name{btw_tool_search_package_info}
\alias{btw_tool_search_package_info}
\alias{btw_tool_cran_package}
\title{Tool: Describe a CRAN package}
\usage{
btw_tool_search_package_info(package_name, `_intent` = NULL)

btw_tool_cran_package(package_name, `_intent` = "")
}
\arguments{
\item{package_name}{The name of a package on CRAN.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
An info sheet about the package.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{btw_tool_search_package_info()} was renamed to
\code{\link[=btw_tool_cran_package]{btw_tool_cran_package()}} in btw 1.2.0.

Describes a CRAN package using \code{\link[pkgsearch:cran_package]{pkgsearch::cran_package()}}.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) withAutoprint(\{ # examplesIf}
cli::cat_line(
  btw_this(pkgsearch::cran_package("anyflights"))
)

\dontshow{\}) # examplesIf}
}
\seealso{
Other cran tools: 
\code{\link{btw_tool_search_packages}()}
}
\concept{cran tools}
\keyword{internal}
