% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-env.R
\name{btw_tool_describe_environment}
\alias{btw_tool_describe_environment}
\title{Tool: Describe an environment}
\usage{
btw_tool_describe_environment(environment = global_env(), items = NULL)
}
\arguments{
\item{environment}{An environment to describe.}

\item{items}{Optional. A character vector of objects in the environment to
describe.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Tool: Describe an environment
}
\seealso{
\code{\link[=btw_this.environment]{btw_this.environment()}}, \code{\link[=btw_register_tools]{btw_register_tools()}}

Other Tools: 
\code{\link{btw_register_tools}()},
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_get_installed_packages}()},
\code{\link{btw_tool_package_docs}}
}
\concept{Tools}
