% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-docs-news.R
\name{btw_tool_docs_package_news}
\alias{btw_tool_docs_package_news}
\title{Tool: Package Release Notes}
\usage{
btw_tool_docs_package_news(package_name, search_term = "")
}
\arguments{
\item{package_name}{The name of the package as a string, e.g. \code{"shiny"}.}

\item{search_term}{A regular expression to search for in the NEWS entries.
If empty, the release notes of the current installed version is included.}
}
\value{
Returns the release notes for the currently installed version of the
package, or the release notes matching the search term.
}
\description{
Include release notes for a package, either the release notes for the most
recent package release or release notes matching a search term.
}
\examples{
# Copy release notes to the clipboard for use in any AI app
btw("@news dplyr", clipboard = FALSE)

btw("@news dplyr join_by", clipboard = FALSE)

if (R.version$major == 4 && R.version$minor > "2.0") {
  # Should find a NEWS entry from R 4.2
  btw("@news R dynamic rd content", clipboard = FALSE)
}

# Tool use by LLMs via ellmer or MCP ----
btw_tool_docs_package_news("dplyr")

btw_tool_docs_package_news("dplyr", "join_by")

}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other Tools: 
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
