% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-data-frame.R
\name{btw_tool_env_describe_data_frame}
\alias{btw_tool_env_describe_data_frame}
\title{Tool: Describe data frame}
\usage{
btw_tool_env_describe_data_frame(
  data_frame,
  format = c("skim", "glimpse", "print", "json"),
  dims = c(5, 100)
)
}
\arguments{
\item{data_frame}{The data frame to describe}

\item{format}{One of \code{"skim"}, \code{"glimpse"}, \code{"print"}, or \code{"json"}.
\itemize{
\item \code{"skim"} is the most information-dense format for describing the data. It
uses and returns the same information as \code{\link[skimr:skim]{skimr::skim()}} but formatting as
a JSON object that describes the dataset.
\item To glimpse the data column-by-column, use \code{"glimpse"}. This is
particularly helpful for getting a sense of data frame column names,
types, and distributions, when pairings of entries in individual rows
aren't particularly important.
\item To just print out the data frame, use \code{print()}.
\item To get a json representation of the data, use \code{"json"}. This is
particularly helpful when the pairings among entries in specific rows
are important to demonstrate.
}}

\item{dims}{The number of rows and columns to show, as a numeric vector of
length two. For example, the default \code{dims = c(5, 100)} shows the first 5
rows and 100 columns, whereas \code{dims = c(Inf, Inf)} would show all of the data.}
}
\value{
A character vector containing a representation of the data frame.
Will error if the named data frame is not found in the environment.
}
\description{
Tool: Describe data frame
}
\examples{
btw_tool_env_describe_data_frame(mtcars)

}
\seealso{
\code{\link[=btw_this.data.frame]{btw_this.data.frame()}}, \code{\link[=btw_tools]{btw_tools()}}

Other Tools: 
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
