% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-environment.R
\name{btw_tool_env_describe_environment}
\alias{btw_tool_env_describe_environment}
\title{Tool: Describe an environment}
\usage{
btw_tool_env_describe_environment(
  items = NULL,
  ...,
  environment = global_env()
)
}
\arguments{
\item{items}{Optional. A character vector of objects in the environment to
describe.}

\item{...}{Ignored.}

\item{environment}{An environment to describe.}
}
\value{
A string describing the environment contents with \verb{#>} prefixing
each object's printed representation.
}
\description{
Tool: Describe an environment
}
\examples{
my_cars <- mtcars[mtcars$mpg > 25, ]
 btw_tool_env_describe_environment("my_cars")

}
\seealso{
\code{\link[=btw_this.environment]{btw_this.environment()}}, \code{\link[=btw_tools]{btw_tools()}}

Other Tools: 
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
