% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-files-list.R
\name{btw_tool_files_list}
\alias{btw_tool_files_list}
\title{Tool: List files}
\usage{
btw_tool_files_list(
  path = NULL,
  type = c("any", "file", "directory"),
  regexp = "",
  `_intent` = ""
)
}
\arguments{
\item{path}{Path to a directory or file for which to get information. The
\code{path} must be in the current working directory. If \code{path} is a directory,
we use \code{\link[fs:dir_ls]{fs::dir_info()}} to list information about files and directories in
\code{path} (use \code{type} to pick only one or the other). If \code{path} is a file, we
show information about that file.}

\item{type}{File type(s) to return, one of \code{"any"} or \code{"file"} or
\code{"directory"}.}

\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns a character table of file information.
}
\description{
Tool: List files
}
\examples{
withr::with_tempdir({
  write.csv(mtcars, "mtcars.csv")

  btw_tool_files_list(type = "file")
})

}
\seealso{
Other files tools: 
\code{\link{btw_tool_files_edit}()},
\code{\link{btw_tool_files_read}()},
\code{\link{btw_tool_files_replace}()},
\code{\link{btw_tool_files_search}()},
\code{\link{btw_tool_files_write}()}
}
\concept{files tools}
