% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-files.R
\name{btw_tool_files_write}
\alias{btw_tool_files_write}
\title{Tool: Write a text file}
\usage{
btw_tool_files_write(path, content, `_intent` = "")
}
\arguments{
\item{path}{Path to the file to write. The \code{path} must be in the current
working directory.}

\item{content}{The text content to write to the file. This should be the
complete content as the file will be overwritten.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns a message confirming the file was written.
}
\description{
Tool: Write a text file
}
\examples{
withr::with_tempdir({
  btw_tool_files_write("example.txt", "Hello\nWorld!")
  readLines("example.txt")
})

}
\seealso{
Other files tools: 
\code{\link{btw_tool_files_edit}()},
\code{\link{btw_tool_files_list}()},
\code{\link{btw_tool_files_read}()},
\code{\link{btw_tool_files_replace}()},
\code{\link{btw_tool_files_search}()}
}
\concept{files tools}
