% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{btw_tool_get_installed_packages}
\alias{btw_tool_get_installed_packages}
\title{Tool: Describe installed packages}
\usage{
btw_tool_get_installed_packages()
}
\value{
Returns a single string describing installed packages as JSON.
}
\description{
Displays the name and title of all installed R packages in JSON format.
}
\examples{
cat(btw_tool_get_installed_packages())

}
\seealso{
\code{\link[=btw_register_tools]{btw_register_tools()}}

Other Tools: 
\code{\link{btw_register_tools}()},
\code{\link{btw_tool_describe_data_frame}()},
\code{\link{btw_tool_describe_environment}()},
\code{\link{btw_tool_package_docs}}
}
\concept{Tools}
