% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-pkg-devtools.R
\name{btw_tool_pkg_document}
\alias{btw_tool_pkg_document}
\title{Tool: Generate package documentation}
\usage{
btw_tool_pkg_document(pkg = ".", `_intent` = "")
}
\arguments{
\item{pkg}{Path to package directory. Defaults to \code{"."}. Must be within
current working directory.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
The output from \code{\link[devtools:document]{devtools::document()}}.
}
\description{
Generate package documentation using \code{\link[devtools:document]{devtools::document()}}. This runs
\pkg{roxygen2} on the package to create/update man pages and \code{NAMESPACE}.
}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_code_search}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_files_write_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_pkg_check}()},
\code{\link{btw_tool_pkg_coverage}()},
\code{\link{btw_tool_pkg_test}()},
\code{\link{btw_tool_run_r}()},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tool_web_read_url}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
