% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-pkg-devtools.R
\name{btw_tool_pkg_test}
\alias{btw_tool_pkg_test}
\title{Tool: Run package tests}
\usage{
btw_tool_pkg_test(pkg = ".", filter = NULL, `_intent` = "")
}
\arguments{
\item{pkg}{Path to package directory. Defaults to '.'. Must be within
current working directory.}

\item{filter}{Optional regex to filter test files. Example: 'helper' matches
'test-helper.R'.}

\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
The output from \code{\link[devtools:test]{devtools::test()}}.
}
\description{
Run package tests using \code{\link[devtools:test]{devtools::test()}}. Optionally filter tests by name
pattern.
}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other pkg tools: 
\code{\link{btw_tool_pkg_check}()},
\code{\link{btw_tool_pkg_coverage}()},
\code{\link{btw_tool_pkg_document}()},
\code{\link{btw_tool_pkg_load_all}()}
}
\concept{pkg tools}
