% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-search-packages.R
\name{btw_tool_search_package_info}
\alias{btw_tool_search_package_info}
\title{Tool: Describe a CRAN package}
\usage{
btw_tool_search_package_info(package_name)
}
\arguments{
\item{package_name}{The name of a package on CRAN.}
}
\value{
An info sheet about the package.
}
\description{
Describes a CRAN package using \code{\link[pkgsearch:cran_package]{pkgsearch::cran_package()}}.
}
\examples{
cli::cat_line(
  btw_this(pkgsearch::cran_package("anyflights"))
)

\dontrun{
btw_tool_search_package_info("anyflights")
}

}
