% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-sessioninfo.R
\name{btw_tool_session_platform_info}
\alias{btw_tool_session_platform_info}
\title{Tool: Describe user's platform}
\usage{
btw_tool_session_platform_info()
}
\value{
Returns a string describing the user's platform.
}
\description{
Describes the R version, operating system, and language and locale settings
for the user's system. When using \code{\link[=btw_client]{btw_client()}} or \code{\link[=btw_app]{btw_app()}}, this
information is automatically included in the system prompt.
}
\examples{
btw_tool_session_platform_info()

}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
