% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-session-info.R
\name{btw_tool_session_platform_info}
\alias{btw_tool_session_platform_info}
\title{Tool: Describe user's platform}
\usage{
btw_tool_session_platform_info(`_intent` = "")
}
\arguments{
\item{_intent}{An optional string describing the intent of the tool use.
When the tool is used by an LLM, the model will use this argument to
explain why it called the tool.}
}
\value{
Returns a string describing the user's platform.
}
\description{
Describes the R version, operating system, and language and locale settings
for the user's system. When using \code{\link[=btw_client]{btw_client()}} or \code{\link[=btw_app]{btw_app()}}, this
information is automatically included in the system prompt.
}
\examples{
btw_tool_session_platform_info()

}
\seealso{
\code{\link[=btw_tools]{btw_tools()}}

Other session tools: 
\code{\link{btw_tool_session_check_package_installed}()},
\code{\link{btw_tool_session_package_info}()}
}
\concept{session tools}
