% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-web.R
\name{btw_tool_web_read_url}
\alias{btw_tool_web_read_url}
\title{Tool: Read a Web Page as Markdown}
\usage{
btw_tool_web_read_url(
  url,
  ...,
  max_wait_for_page_load_s = getOption("btw.max_wait_for_page_load_s", 10)
)
}
\arguments{
\item{url}{The URL of the web page to read.}

\item{...}{Ignored, for future features.}

\item{max_wait_for_page_load_s}{Maximum time to wait for the page to load, in
seconds. Can be set globally using the \code{btw.max_wait_for_page_load_s}
option.}
}
\description{
Tool: Read a Web Page as Markdown
}
\examples{
\dontshow{if (rlang::is_installed("chromote") && rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
btw_tool_web_read_url("https://www.r-project.org/")
btw_tool_web_read_url(
  "https://posit.co/blog/easy-tool-calls-with-ellmer-and-chatlas/"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_files_write_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tools}()}
}
\concept{Tools}
