% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{btw_tools}
\alias{btw_tools}
\title{Tools: Register tools from btw}
\usage{
btw_tools(...)
}
\arguments{
\item{...}{Optional names of tools or tool groups to include when registering
tools. By default all btw tools are included. For example, use \code{tools = "docs"} to include only the documentation related tools, or \code{tools = c("env", "docs", "session")}, etc.}
}
\value{
Registers the tools with \code{chat}, updating the \code{chat} object in
place. The \code{chat} input is returned invisibly.
}
\description{
The \code{btw_tools()} function provides a list of tools that can be registered
with an ellmer chat via \code{chat$set_tools()} that allow the chat to
interface with your computational environment. Chats returned by
this function have access to the tools:

\verb{r .docs_list_tools()}
}
\examples{
# requires an ANTHROPIC_API_KEY
\dontrun{
ch <- ellmer::chat_anthropic()

# register all of the available tools
ch$set_tools(btw_tools())

# or register only the tools related to fetching documentation
ch$set_tools(btw_tools("docs"))

# ensure that the current tools persist
ch$set_tools(c(ch$get_tools(), btw_tools()))
}

}
\seealso{
Other Tools: 
\code{\link{btw_tool_docs_package_news}()},
\code{\link{btw_tool_env_describe_data_frame}()},
\code{\link{btw_tool_env_describe_environment}()},
\code{\link{btw_tool_files_list_files}()},
\code{\link{btw_tool_files_read_text_file}()},
\code{\link{btw_tool_files_write_text_file}()},
\code{\link{btw_tool_ide_read_current_editor}()},
\code{\link{btw_tool_package_docs}},
\code{\link{btw_tool_search_packages}()},
\code{\link{btw_tool_session_package_info}()},
\code{\link{btw_tool_session_platform_info}()},
\code{\link{btw_tool_web_read_url}()}
}
\concept{Tools}
