% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkg.R
\name{get_pkg}
\alias{get_pkg}
\alias{get_package_help}
\alias{get_help_page}
\title{Describe R package documentation}
\usage{
get_package_help(package_name)

get_help_page(package_name, topic)
}
\arguments{
\item{package_name}{The name of the package as a string, e.g. \code{"shiny"}.}

\item{topic}{The \code{topic_id} or \code{alias} of the help page, e.g. \code{"withProgress"}
or \code{"incProgress"}. Find \code{topic_id}s or \code{alias}es using \code{get_package_help()}.}
}
\value{
\itemize{
\item \code{get_package_help()} returns the \code{topic_id}, \code{title}, and \code{aliases} fields
for every topic in a package's documentation as a json-formatted string.
\item \code{get_help_page()} return the help-page for a package topic as a string.
}
}
\description{
These functions describe package documentation in plain text:
}
\section{See Also}{

\itemize{
\item \code{\link[=register_btw_tools]{register_btw_tools()}} registers this function as a tool for LLMs to call.
\item Other \verb{get_*()} functions: \code{\link[=get_data_frame]{get_data_frame()}}, \code{\link[=get_environment]{get_environment()}}, \code{\link[=get_installed_packages]{get_installed_packages()}}, \code{\link[=get_package_help]{get_package_help()}}, \code{\link[=get_help_page]{get_help_page()}}
}
}

\examples{
cat(get_package_help("btw"))

cat(get_help_page("btw", "btw"))

}
