% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_btw_tools}
\alias{register_btw_tools}
\title{Register tools from btw}
\usage{
register_btw_tools(chat)
}
\arguments{
\item{chat}{An ellmer \code{Chat} object.}
}
\value{
The chat object with tools registered.
}
\description{
The \code{register_btw_tools()} function equips an ellmer chat to interface with
your computational environment. Chats returned by this function have access
to the tools \code{\link[=btw_show_data_frame_structure]{btw_show_data_frame_structure()}}, \code{\link[=btw_list_and_describe_environment]{btw_list_and_describe_environment()}}, \code{\link[=get_installed_packages]{get_installed_packages()}}, \code{\link[=get_package_help]{get_package_help()}}, \code{\link[=get_package_vignette]{get_package_vignette()}}, \code{\link[=get_package_vignettes]{get_package_vignettes()}}, \code{\link[=get_help_page]{get_help_page()}}.
}
\examples{
# requires an ANTHROPIC_API_KEY
\dontrun{
ch <- ellmer::chat_claude()

register_btw_tools(ch)
}
}
